+++ Tests/LirResourcesClientTests/UnitTest1.cs
@@ -6,7 +6,7 @@ public class LirResourcesClientIntegrationUnitTests
 {
     private readonly LirResourcesClient _client;
 
-    private readonly string _apiKey;
+    private readonly string? _apiKey;
     
     public LirResourcesClientIntegrationUnitTests()
     {
@@ -20,7 +20,7 @@ public class LirResourcesClientIntegrationUnitTests
         _client.Debug = true;
     }
 
-    private string GetApiKey()
+    private string? GetApiKey()
     {
         return _apiKey;
     }
@@ -28,7 +28,7 @@ public class LirResourcesClientIntegrationUnitTests
     [Fact]
     public async Task RequestIpv4()
     {
-        var ipv4 = await _client.GetIpv4(GetApiKey());
+        var ipv4 = await _client.GetIpv4(GetApiKey(), CancellationToken.None);
         Assert.True(ipv4.Ipv4Allocations.Any(), "No Ipv4 PA networks from LIR API.");
         Assert.True(ipv4.Ipv4Assignments.Any(), "No Ipv4 PI networks from LIR API.");
     }
@@ -36,14 +36,14 @@ public class LirResourcesClientIntegrationUnitTests
     [Fact]
     public async Task RequestIpv6()
     {
-        var ipv6 = await _client.GetIpv6(GetApiKey());
+        var ipv6 = await _client.GetIpv6(GetApiKey(), CancellationToken.None);
         Assert.True(ipv6.Ipv6Allocations.Any(), "No Ipv6 networks from LIR API.");
     }
     
     [Fact]
     public async Task RequestAll()
     {
-        var resources = await _client.GetResources(GetApiKey());
+        var resources = await _client.GetResources(GetApiKey(), CancellationToken.None);
         
         Assert.True(resources.Ipv4Allocations.Any(), "No Ipv4 PA networks from LIR API.");
         Assert.True(resources.Ipv4Assignments.Any(), "No Ipv4 PI networks from LIR API.");
