+++ src/ClientsRipe/LirResources/LirResourcesClient.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using ClientsRipe.LirResources.Models;
 using ClientsRipe.RpkiClient.Models;
@@ -9,17 +10,10 @@ using RestSharp;
 
 namespace ClientsRipe.LirResources;
 
-public interface ILirResourcesClient
-{
-    public bool Debug { get; set; } 
-       
-    public Task<LirResourcesReply> GetResources(string apiKey);
-}
-
 public class LirResourcesClient : ILirResourcesClient
 {
     private readonly string _baseUrl;
-        
+
     public LirResourcesClient(ILirResourcesLocation ripeRpkiLocation)
     {
         _baseUrl = ripeRpkiLocation.Url;
@@ -28,7 +22,7 @@ public class LirResourcesClient : ILirResourcesClient
     private RestClient GetClient(string apiKey)
     {
         RestClient client;
-            
+
         if (Debug)
         {
             var options = new RestClientOptions(_baseUrl)
@@ -45,55 +39,54 @@ public class LirResourcesClient : ILirResourcesClient
         }
 
         client.AddDefaultHeader("ncc-api-authorization", apiKey);
-            
+
         return client;
     }
 
     public bool Debug { get; set; }
-    
-    public async Task<LirResourcesReply> GetAll(string apiKey)
+
+    public async Task<LirResourcesReply> GetAll(string apiKey, CancellationToken cancellationToken)
     {
-        return await RequestResources("", apiKey);
-    }
-    
-    
-    public async Task<LirResourcesReply> Get(string apiKey)
-    {
-        return await RequestResources("", apiKey);
-    }
-    
-    public async Task<LirResourcesReply> GetAsn(string apiKey)
-    {
-        return await RequestResources("asn", apiKey);
-    }
-    
-    public async Task<LirResourcesReply> GetIpv4(string apiKey)
-    {
-        return await RequestResources("ipv4", apiKey);
+        return await RequestResources("", apiKey, cancellationToken);
     }
 
-    public async Task<LirResourcesReply> GetIpv6(string apiKey)
+    public async Task<LirResourcesReply> Get(string apiKey, CancellationToken cancellationToken)
     {
-        return await RequestResources("ipv6", apiKey);
+        return await RequestResources("", apiKey, cancellationToken);
     }
 
-    public async Task<LirResourcesReply> GetResources(string apiKey)
+    public async Task<LirResourcesReply> GetAsn(string apiKey, CancellationToken cancellationToken)
     {
-        var resources = await GetAll(apiKey);
+        return await RequestResources("asn", apiKey, cancellationToken);
+    }
+
+    public async Task<LirResourcesReply> GetIpv4(string apiKey, CancellationToken cancellationToken)
+    {
+        return await RequestResources("ipv4", apiKey, cancellationToken);
+    }
+
+    public async Task<LirResourcesReply> GetIpv6(string apiKey, CancellationToken cancellationToken)
+    {
+        return await RequestResources("ipv6", apiKey, cancellationToken);
+    }
+
+    public async Task<LirResourcesReply> GetResources(string apiKey, CancellationToken cancellationToken)
+    {
+        var resources = await GetAll(apiKey, cancellationToken);
         //TODO: Not implemented
         resources.Ipv6Assignments = null;
-        
+
         return resources;
     }
-    
-    
-    protected virtual async Task<LirResourcesReply> RequestResources(string resource, string apiKey)
+
+
+    protected virtual async Task<LirResourcesReply> RequestResources(string resource, string apiKey, CancellationToken cancellationToken)
     {
         if (string.IsNullOrEmpty(apiKey))
             throw new ArgumentException("API key not provided.", nameof(apiKey));
-            
+
         var request = new RestRequest(resource, Method.Get);
-        
+
         request.AddParameter("format", "json");
         request.AddParameter("jsonCallback", "?");
 
@@ -101,7 +94,7 @@ public class LirResourcesClient : ILirResourcesClient
 
         try
         {
-            var reply = await client.ExecuteAsync<LirResourcesReply>(request);
+            var reply = await client.ExecuteAsync<LirResourcesReply>(request, cancellationToken);
             if (!reply.IsSuccessful)
                 throw new Exception(reply.Content);
 
@@ -112,6 +105,6 @@ public class LirResourcesClient : ILirResourcesClient
         {
             Console.WriteLine(e);
             throw;
-        }        
+        }
     }
 }
