+++ src/ClientsRipe/RipeClient/Exceptions/RipeClientAuthPasswordException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace ClientsRipe
+{
+    public class RipeClientAuthPasswordException : Exception
+    {
+        public RipeClientAuthPasswordException(string message) : base(message)
+        {
+            
+        }
+    }
+}
