+++ src/ClientsRipe/RipeClient/Exceptions/RipeClientConflictException.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace ClientsRipe
+{
+    public class RipeClientConflictException: Exception
+    {
+        public RipeClientConflictException(string content)
+        {
+            Content = content;
+        }
+        
+        public string Content { get; }
+    }
+}
