+++ src/ClientsRipe/RipeClient/Exceptions/RipeClientException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ClientsRipe
+{
+    public class RipeClientException : Exception
+    {
+        public RipeClientException(string message) : base(message)
+        {
+        }
+    }
+}
