+++ src/ClientsRipe/RipeClient/Exceptions/RipeClientNotFoundException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ClientsRipe
+{
+    public class RipeClientNotFoundException : Exception
+    {
+        public RipeClientNotFoundException(string message) : base(message)
+        {
+        }
+    }
+}
