+++ src/ClientsRipe/RipeClient/Exceptions/RipeSearchNotFoundException.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace ClientsRipe
+{
+    public class RipeSearchNotFoundException : Exception
+    {
+        public RipeSearchNotFoundException() : base() { }
+        public RipeSearchNotFoundException(string message) : base(message)  { }
+    }
+}
