+++ src/ClientsRipe/RipeClient/IRipeClientAuth.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace ClientsRipe
+{
+    public interface IRipeClientAuth
+    {
+        Task<string> GetSecret();
+    }
+}
