+++ src/ClientsRipe/RipeClient/IRipeLocation.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientsRipe
+{
+    public interface IRipeLocation
+    {
+        public string Url { get; set; }
+    }
+}
