+++ src/ClientsRipe/RipeClient/RipeClientAuthAnonymous.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+
+namespace ClientsRipe
+{
+    public class RipeClientAuthAnonymous : IRipeClientAuth
+    {
+        public Task<string> GetSecret()
+        {
+            return Task.FromResult("");
+        }
+    } 
+}
