+++ src/ClientsRipe/RipeClient/RipeClientAuthPassword.cs
@@ -1,0 +1,19 @@
+﻿using System.Threading.Tasks;
+
+namespace ClientsRipe
+{
+    public class RipeClientAuthPassword : IRipeClientAuth
+    {
+        private readonly string _password;
+
+        public RipeClientAuthPassword(string password)
+        {
+            _password = password;
+        }
+        
+        public Task<string> GetSecret()
+        {
+            return Task.FromResult(_password);
+        }
+    }
+}
