+++ src/ClientsRipe/RipeClient/RipeNonSecureLocation.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientsRipe
+{
+    public class RipeNonSecureLocation : IRipeLocation
+    {
+        public string Url { get; set; } = "http://rest.db.ripe.net/";
+    }
+}
