+++ src/ClientsRipe/RipeClient/RipeSecureLocation.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientsRipe
+{
+    public class RipeSecureLocation : IRipeLocation
+    {
+        public string Url { get; set; } = "https://rest.db.ripe.net/";
+    }
+}
