--- src/ClientsRipe/RipeLocation.cs
@@ -1,17 +1,0 @@
-﻿namespace ClientsRipe
-{
-    public interface IRipeLocation
-    {
-        public string Url { get; set; }
-    }
-
-    public class RipeNonSecureLocation : IRipeLocation
-    {
-        public string Url { get; set; } = "http://rest.db.ripe.net/";
-    }
-
-    public class RipeSecureLocation : IRipeLocation
-    {
-        public string Url { get; set; } = "https://rest.db.ripe.net/";
-    }
-}
