+++ src/ClientsRipe/RpkiClient/IRipeRPKIRoute.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientsRpki
+{
+    public interface IRipeRPKIRoute
+    {
+        
+    }
+}
