+++ src/ClientsRipe/RpkiClient/IRipeRpkiSettingsManager.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientsRpki
+{
+    public interface IRipeRpkiSettingsManager
+    {
+        public RpkiSettings LoadSettings();
+    }
+}
